Adds calculator class and breaks up sales analyst tests.
